pyproject.toml
[tool.poetry]
name = "food-order-app"
version = "0.1.0"
description = ""
authors = ["spacecat1717 <manyam245@gmail.com>"]
readme = "README.md"
packages = [{include = "food_order_app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.12"}
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
pylint = "^2.17.4"
email-validator = "^2.0.0.post2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
